# arm64: reduces the struct size of instructions