… internal/engine/wazevo/backend/isa/arm64/instr_encoding_test.go
… internal/engine/wazevo/backend/isa/arm64/instr_test.go
… internal/engine/wazevo/backend/isa/arm64/lower_instr_test.go
… internal/engine/wazevo/backend/isa/arm64/lower_mem.go
… internal/engine/wazevo/backend/isa/arm64/lower_mem_test.go
… internal/engine/wazevo/backend/isa/arm64/machine.go
… internal/engine/wazevo/backend/isa/arm64/machine_pro_epi_logue.go
… internal/engine/wazevo/backend/isa/arm64/machine_test.go